Use a selector to get the selected reward's required todos instead of querying the database
Use a selector to get the selected reward's required todos instead of querying the database with /todos-for-selection every time the selectedRewardId or requirements change. Instead of using the associated reward that was generated by a todos and requirements join in the database, this uses a boolean value of if the todoId is in the selected reward's requirements.