Fix cypress test and add testid to check in ToggleButton
Selectors were not digging deep enough to get to the check. The selector has to also select the button within .toggleButton. Instead I added a testid to the check in ToggleButton so they can be found easier. Using .find() prevents needing to dig all the way down to the exact parent of the check.